+++ MaRTA/src/structs/job.h
@@ -6,22 +6,22 @@
 
 typedef struct {
 	char nodeName[25];
-	uint32_t numBlock;
+	uint16_t numBlock;
 } t_copy;
 
 typedef struct {
-	uint32_t id;
+	uint16_t id;
 	t_list *copies;
 	char *nodeName;
 	char *nodeIP;
 	uint16_t nodePort;
-	uint32_t numBlock;
+	uint16_t numBlock;
 	char tempResultName[60];
 	bool done;
 } t_map;
 
 typedef struct {
-	uint32_t originMap;
+	uint16_t originMap;
 	char *nodeIP;
 	uint16_t nodePort;
 	char tempName[60];
@@ -42,7 +42,7 @@ typedef struct {
 } t_file;
 
 typedef struct {
-	uint32_t id;
+	uint16_t id;
 	bool combiner;
 	t_list *files;
 	t_list *maps;
@@ -50,10 +50,10 @@ typedef struct {
 	t_reduce *finalReduce;
 } t_job;
 
-t_copy *CreateCopy(char *nodeName, uint32_t numBlock);
+t_copy *CreateCopy(char *nodeName, uint16_t numBlock);
 t_file *CreateFile(char *path);
-t_job *CreateJob(uint32_t id, bool combiner);
+t_job *CreateJob(uint16_t id, bool combiner);
 void freeJob(t_job *job);
-bool isMap(t_map *map, uint32_t idMap);
+bool isMap(t_map *map, uint16_t idMap);
 
 #endif
